Moved elastic files to an elastic specific folder
