+++ lib/elastic/helpers.js
