+++ lib/elastic/mappings/pages.yml
