+++ lib/services/page-list.js
@@ -3,8 +3,8 @@
 const bluebird = require('bluebird'),
   _ = require('lodash'),
   sites = require('./sites'),
-  filters = require('../utils/db-filters'),
-  elasticHelpers = require('../utils/elastic-helpers'),
+  filters = require('../elastic/db-filters'),
+  elasticHelpers = require('../elastic/helpers'),
   search = require('./search'),
   files = require('../files'),
   path = require('path'),
@@ -13,7 +13,7 @@ const bluebird = require('bluebird'),
 
 function updatePageList(mappings) {
   return function (batchOps) {
-    return bluebird.all(_.map(_.pick(mappings, ['editable-pages']), function (types, indexName) {
+    return bluebird.all(_.map(_.pick(mappings, ['pages']), function (types, indexName) {
       return bluebird.all(_.map(types, function (mapping, typeName) {
         var ops = _.cloneDeep(batchOps);
