+++ lib/services/search.js
@@ -3,7 +3,7 @@
 const elastic = require('elasticsearch'),
   bluebird = require('bluebird'),
   _ = require('lodash'),
-  elasticHelpers = require('../utils/elastic-helpers'),
+  elasticHelpers = require('../elastic/helpers'),
   db = require('./db'),
   pageList = require('./page-list'),
   files = require('../files'),
@@ -400,7 +400,7 @@ function retrieveMappingData(dirPath) {
  * @return {Promise}
  */
 function createInternalIndices() {
-  var dirPath = path.resolve(__dirname, '../mappings'),
+  var dirPath = path.resolve(__dirname, '../elastic/mappings'),
     internalMappings = retrieveMappingData(dirPath);
 
   // TODO: check for folder and index file inside of Sites
@@ -409,8 +409,6 @@ function createInternalIndices() {
     .then(function() {
       // Listen to events internally
       db.on('batch', pageList.updatePageList(internalMappings));
-      // Listen to events in the Clay instance
-      // Resolve the promise
       return bluebird.resolve();
     });
 }
@@ -436,7 +434,7 @@ module.exports.retrieveMappingData = retrieveMappingData;
 module.exports.validateIndices = validateIndices;
 
 // Filters for Elastic ops
-module.exports.filters = require('../utils/db-filters');
+module.exports.filters = require('../elastic/db-filters');
 // Elastic Helpers
 module.exports.convertRedisBatchtoElasticBatch = elasticHelpers.convertRedisBatchtoElasticBatch;
 module.exports.normalizeOpValuesWithMapping = elasticHelpers.normalizeOpValuesWithMapping;
